Add scroll-padding-top to style.css
When the "Hall of Fame" button at the top is pressed, the page scrolls down and leaves no space above the "Hall of fame" heading (i.e. between the top edge of the heading text and the top edge of the viewport). With this property the page scrolls down and leaves 2rem space above the heading.